Enable -subfps for time based subtitles, too.
Please update the DOCS for -subfps:
"If you have time-based subtitles like subviewer that can't be synced by -subdelay
try specifying -subfps with the framerate of the movie the time-based sub was
created for, eg. -subfps 25 if you use subs from a 25 fps movie with a movie
using a different framerate. If you need to guess -subfps, try 25, 23.98 and 29.97"


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@7611 b3059339-0415-0410-9bf9-f77b7e298cf2