major reindentation of x11 code try #2
note that this is plain ident output, i didnt tweak it by
hand like the last attempt.

if anyone is interested in the indent profile i used, just drop me a mail.

please contact me on irc on how to send me my share of cola,
but be aware that i will only accept swiss or german cola, as the japanese is
way to sweet :)


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@12583 b3059339-0415-0410-9bf9-f77b7e298cf2