Currently vbeGetProtModeInfo call the 0x4f0a function of int 10h the get a simple 32 bits protected mode interface to some VESA functions. This protected mode interface is interesting because it's quicker than the raw int 10h interface. Unfortunatly, begining with VBE 3.0, the 0x4f0a function is optional, and some video cards don't implement it (3dfx, intel 845/855/865...). This protected mode interface is then only used in vbeSetWindow and vbeSetDisplayStart :  - vbeSetWindow already implement an alternative methode if protected mode interface is not available.  - vbeSetDisplayStart also contain an alternative implementation, but this one is disabled with a #if 0. I don't exactly know why because it works well !
So currently, cards which don't have the 0x4f0a function are not
supported. This patch correct this.
 - vbeGetProtModeInfo failure is not fatal.
 - vbeSetDisplayStart has it's alternative implementation reenabled.
   it's used only with cards which don't have the 0x4f0a function
   so this won't make any difference for cards which were already
   working.

This patch also make the failure of vbeGetModeInfo not fatal. The
VBE 3.0 standard state that GetModeInfo can fail with some mode
which are listed as supported if the mode can't be used in the
current situation (not enough video memory for example). So a
failure of vbeGetModeInfo don't mean that other modes won't work
and should really not be fatal.

patch by Aurelien Jacobs <aurel@gnuage.org>


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@13569 b3059339-0415-0410-9bf9-f77b7e298cf2