honor decoder's/filter's decision to remove frames when using -noskip. this may go against the original intention of the vf layer, but it's how all the filters that drop frames have been written to work, so it's now the de-facto standard.
without this patch, -noskip will result in tons of duplicate frames
(either soft or hard duplicates) and a/v desync whenever decimation,
ivtc, etc. is used. even with this patch -noskip is still a bad idea
for most of these purposes, but it will work reliably with filmdint,
framestep, and some other filters with fixed in:out ratios as long as
the right -ofps value is used. without this patch, there is no hope of
-noskip working with frame-dropping filters.

(this patch was previously committed erroneously as part of another
change, then reversed. it is now being committed again.)


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@16830 b3059339-0415-0410-9bf9-f77b7e298cf2