Always enable largefile support by default
The largefile configure option was disabled by default, but the
enabled-by-default dvdread and dvdcss options force in on (dvdnav
code also tries to enable it, but sets the variable too late so that
is has no effect). Change configure to enable largefile support
independently of other options unless explicitly disabled.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@25400 b3059339-0415-0410-9bf9-f77b7e298cf2