Fix memory leak in appInitStruct().
The NumberOf{Main,Bar,Menu}Items members actually don't hold the
number of items but the last index used. Therefore, the last item
won't get freed so far.

To set the correct "number" for the global external variable prior
to using it, the function appResetStruct() has been added and is used
in appInitStruct() as well.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@33014 b3059339-0415-0410-9bf9-f77b7e298cf2