Remove function Normalize().
This function changed ARGB data as if it had been RGB24 data (not quite
correctly) converted to ARGB with all colors transparent.

Instead now, set the alpha channel for RGB24 data and leave ARGB data
untouched.

For X11, the alpha channel is disregarded, so both approaches are equal,
but the new one is more intelligible as we get correct ARGB PNG data.

For legacy reasons, all kind of fuchsia/magenta must be treated as
transparent, because some skins are using at least both full opaque and
full transparent fuchsia/magenta for transparency.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@33622 b3059339-0415-0410-9bf9-f77b7e298cf2