configure: simplify -lm handling
Add -lm to extra_ldflags like all other linker flags instead of having a
separate variable for it. This simplifies the linker flag handling and
should hopefully address Bugzilla #984.

Note that this may well have unforeseen sideeffects on strange platforms
that I cannot test, but this is clearly the way forward, so any problems
will have to fixed as they appear.

git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@33885 b3059339-0415-0410-9bf9-f77b7e298cf2