libmpcodec: add vf_lavfi.
This filter wraps a complete libavfilter filter graph.

As the API of libavfilter is not completely stable yet, the filter is not
enabled by default, so as not to let mplayer unbuildable.

Some strange behaviours may appear due to the very different model of buffer
allocation between mplayer and lavfi.

git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@34190 b3059339-0415-0410-9bf9-f77b7e298cf2