ao_alsa: drop check for sys/asoundlib.h and ALSA version
The ALSA output driver uses some parts of the ALSA API that were introduced
after the switch to alsa/asoundlib.h, so checking for the deprecated header
location is pointless as it has no chance of working anyway.

Also, now that both ALSA output and input work with the 0.9 and 1.0 series,
there is no need to keep checking for the ALSA version.

git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@34347 b3059339-0415-0410-9bf9-f77b7e298cf2