mp_property_deinterlace: Signalize control() failure
Always having to do a GET after a SET to check if the state was
really changed, seems not very friendly from the perspective of
the client application that changes the property.

Return M_PROPERTY_UNAVAILABLE to signalize that the control() was
not positively consumed by the filter/vo chain.

Based on a patch by Vicente Sendra >visenri yahoo.es<



git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@35267 b3059339-0415-0410-9bf9-f77b7e298cf2