build/configure: Move addition of extra CFLAGS further down
The test is know made on cc_vendor after we correctly identified
the compiler. Before we were just assuming if the program is named
cc it is GCC.

For the record:
The condition may still/or in future be wrong for some compilers on
darwin and I doubt the flags are really necessary nowadays.

Patch by Xidorn Quan >quanxunzhen gmail.com<



git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@35276 b3059339-0415-0410-9bf9-f77b7e298cf2