Don't set "normal size" check mark in menu by default.
The "Normal size" item in the menu is currently marked by default if the
video is neither double nor half size, and even if the video window size
isn't the exact video size.

To avoid confusion, only set the check mark, if window and video are the
same size, and mark nothing if not.

Based on a patch by Hans-Dieter Kosch, hdkosch kabelbw de.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@35780 b3059339-0415-0410-9bf9-f77b7e298cf2