Keep GUI history entries in order.
Currently, if an entry is to be added to the GUI history and this entry
is already existing, it is swapped with the top entry. This breaks
history sequence and is confusing.

Move this entry to the top while leaving the other ones in their order.

Patch by Hans-Dieter Kosch, hdkosch kabelbw de.


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@35941 b3059339-0415-0410-9bf9-f77b7e298cf2