Add a function converting into locale encoding.
This is for compatibility reasons, because the content of the history
file will change from UTF-8 encoded filenames to filenames in their
original byte sequence.

(The cfg-old.c file is meant to save the user the effort of manually
changing any settings or options that have become obsolete. The file
will be subject to removal some day. It will be included where
necessary.)


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@35965 b3059339-0415-0410-9bf9-f77b7e298cf2