Fix issues with glibc on 32 bit.
Some glibc versions e.g. from RedHat will cause
incorrect behaviour when a certain magic symbol
is not encountered.
This is just abhorrent and simply unacceptable,
breaking well-behaving programs to the benefit
of ancient programs, but let's try to make it
work anyway (and hope they won't repeat the insanity).

git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@36346 b3059339-0415-0410-9bf9-f77b7e298cf2